requirements.txt
# ── LLM backends ──────────────────────────────────────────────────────
openai>=1.30.0             # OpenAI GPT-3.5-turbo / GPT-4
anthropic>=0.25.0          # Claude (optional alternative backend)
requests>=2.31.0           # HTTP forwarding in patch proxy

# ── Config & utilities ────────────────────────────────────────────────
python-dotenv>=1.0.0       # loads .env automatically
pyyaml>=6.0                # YAML config files

# ── Patch Proxy ───────────────────────────────────────────────────────
fastapi>=0.111.0           # proxy server framework
uvicorn>=0.29.0            # ASGI server for FastAPI

# ── Testing ───────────────────────────────────────────────────────────
pytest>=8.0

# ── Optional: Garak scanner ──────────────────────────────────────────
# garak>=0.9
